refactor(actions-bar): move audio-device-selector to a different component
